Generate and save an allele frequency distribution plot

This function creates a faceted bar plot showing the frequency of each allele length for Day 0 vs. Day of Failure samples. It can generate plots for a single site or an overall plot for all sites combined.

Usage

generate_allele_frequency_plot(
  raw_data_df,
  site_name = NULL,
  output_folder,
  plots_per_row = 4,
  plot_width = 12,
  plot_height = 10
)

Arguments

raw_data_df

A data frame containing raw genotyping data. For a site-specific plot, this should be filtered for one site. For an overall plot, it should contain data from all sites.

site_name

The name of the site for the plot title and filename. If NULL (the default), an "Overall" plot is generated.

output_folder

The directory where the plot PDF will be saved.

plots_per_row

The number of marker plots to display per row.

plot_width

The width of the output PDF file.

plot_height

The height of the output PDF file.